refactor overview to fit new routing scheme
