+++ lib/routes.js
@@ -74,8 +74,6 @@ routes.cleanAsPath = asPath => asPath.split('?')[0].split('#')[0]
 //  '/:year(\\d{4})/:month(\\d{2})/:day(\\d{2})/:slug/:suffix(diskussion)?'
 //)
 //.add('section', '/:slug', 'article')
-
-// TODO: refactor into 2018/2019/2020
 //.add('overview', '/:year(\\d{4})/:interval(wochen)?')
 
 module.exports = routes
